Serious Memory problems in graphics
- Fixed graphics
- Compute not getting collisons
- fixed ssbo limits
- reran tests
- Fixed plots and tables
- Updated article
	fix sequential vs random
	start detailed outline with paragraghs, bullets, and puropse

+++ vulkan/src/VulkanObj/MemStats.cpp
@@ -4,7 +4,7 @@
 void MemStats(VulkanObj* vulkanObj)
 {
 
-	mout << "Total Memory Allocated:" << vulkanObj->m_TotalMemoryBytes << ende;
+	mout << "Total Memory Allocated:" << uint32_t(vulkanObj->m_TotalMemoryBytes) << ende;
 	VmaTotalStatistics memuse;
 	vmaCalculateStatistics(vulkanObj->m_vmaAllocator, &memuse);
